gsev: set tab_size to 4 on Go files and lazily run callbacks
ST4 sometimes sets the tab_size of Go code to 2. This changes the EV
class to enforce a tab_size of 4 for Go source code.

This commit also changes the EV class to only schedule callbacks when
the code is Go (this should reduce its performance impact on non-Go
source code).

This commit also adds mypy type annotations to gsev.py.